Configure application class and fix context injection in tests
This commit introduces an `EqualizerApplication` class and configures it in `AndroidManifest.xml` by adding the `android:name=".EqualizerApplication"` attribute to the application tag.

Additionally, it fixes the context injection in the `AudioServiceTest` by setting the `_context` field of the `AudioService` instance directly instead of passing it to the constructor. This aligns with the `AudioService` class being modified to initialize `_context` with `this` as a default value.